[관리자] 메뉴 추가 기능

+++ public_project/src/main/java/com/kr/pub/dto/ImageDTO.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class ImageDTO {
 	private String imgId;
 	private Date regDate;
-	private int length;
+	private long length;
 	private String imgNameReal;
 	private char deleted;
 	private String contentType;
